better memory management
add init functions for device and driver. It will make it easy to
claim memory. It is much clear now, who is the onwner of the memeory.
For example, cj_dev will has ownership of its memory and programmers
use it don't need to worry life cycle of memory inside of it.